refacted aws func, moved to common
