+++ pkg/common/utils.go
@@ -4,6 +4,9 @@ import (
 	"bufio"
 	"bytes"
 	"crypto/rand"
+	"encoding/base32"
+	"encoding/binary"
+	"fmt"
 	"io"
 	"math/big"
 	"strings"
@@ -63,4 +66,28 @@ func RandomID(length int) string {
 	}
 
 	return string(b)
+}
+
+func GetAccountNumFromAWSID(AWSID string) (string, error) {
+	// Function to get the account number from an AWS ID (no verification required)
+	// Source: https://medium.com/@TalBeerySec/a-short-note-on-aws-key-id-f88cc4317489
+	if len(AWSID) < 4 {
+		return "", fmt.Errorf("AWSID is too short")
+	}
+	trimmed_AWSID := AWSID[4:]
+	decodedBytes, err := base32.StdEncoding.WithPadding(base32.NoPadding).DecodeString(strings.ToUpper(trimmed_AWSID))
+	if err != nil {
+		return "", err
+	}
+
+	if len(decodedBytes) < 6 {
+		return "", fmt.Errorf("Decoded AWSID is too short")
+	}
+
+	data := make([]byte, 8)
+	copy(data[2:], decodedBytes[0:6])
+	z := binary.BigEndian.Uint64(data)
+	const mask uint64 = 0x7fffffffff80
+	account_num := (z & mask) >> 7
+	return fmt.Sprintf("%012d", account_num), nil
 }
